+++ source/UI/UIObject.cpp
@@ -4,7 +4,7 @@
 namespace UIO
 {
 
-	Entity ui_hp_bar(f32 x, f32 y, f32 width, f32 height, f32 rotation, s8 z)
+	Entity ui_hp_bar(f32 x, f32 y, f32 width, f32 height, f32 rotation, s32 z)
 
 	{
 
@@ -22,7 +22,7 @@ namespace UIO
 		return id;
 	}
 
-	Entity ui_stamina_bar(f32 x, f32 y, f32 width, f32 height, f32 rotation, s8 z)
+	Entity ui_stamina_bar(f32 x, f32 y, f32 width, f32 height, f32 rotation, s32 z)
 	{
 		Entity id = ecs.createEntity();
 		//default player values
@@ -68,7 +68,7 @@ namespace UIO
 
 	}
 
-	Entity ui_button(f32 x, f32 y, f32 width, f32 height, f32 rotation, s8 z, std::function<void()> func)
+	Entity ui_button(f32 x, f32 y, f32 width, f32 height, f32 rotation, s32 z, std::function<void()> func)
 
 	{
 		Entity id = ecs.createEntity();
@@ -90,7 +90,7 @@ namespace UIO
 		return id;
 	}
 
-	Entity ui_button_texture(AEGfxTexture* texture, f32 x, f32 y, f32 width, f32 height, f32 rotation, s8 z, std::function<void()> func)
+	Entity ui_button_texture(AEGfxTexture* texture, f32 x, f32 y, f32 width, f32 height, f32 rotation, s32 z, std::function<void()> func)
 	{
 		Entity id = ecs.createEntity();
 		//default player values
@@ -112,7 +112,7 @@ namespace UIO
 		return id;
 	}
 
-	Entity ui_text(f32 x, f32 y, f32 width, f32 height, f32 rotation, s8 z, const char* a)
+	Entity ui_text(f32 x, f32 y, f32 width, f32 height, f32 rotation, s32 z, const char* a)
 
 	{
 		Entity id = ecs.createEntity();
@@ -142,7 +142,7 @@ namespace UIO
 		return id;
 	}
 
-	Entity ui_blank_solid_corner(f32 x, f32 y, f32 width, f32 height, f32 rotation, s8 z, f32 r, f32 g, f32 b, f32 a)
+	Entity ui_blank_solid_corner(f32 x, f32 y, f32 width, f32 height, f32 rotation, s32 z, f32 r, f32 g, f32 b, f32 a)
 	{
 		Entity id = ecs.createEntity();
 		//default player values
@@ -156,7 +156,7 @@ namespace UIO
 		return id;
 	}
 
-	Entity ui_blank_solid_center(f32 x, f32 y, f32 width, f32 height, f32 rotation, s8 z, f32 r, f32 g, f32 b, f32 a)
+	Entity ui_blank_solid_center(f32 x, f32 y, f32 width, f32 height, f32 rotation, s32 z, f32 r, f32 g, f32 b, f32 a)
 	{
 		Entity id = ecs.createEntity();
 		//default player values
@@ -170,7 +170,7 @@ namespace UIO
 		return id;
 	}
 
-	Entity ui_blank_texture(AEGfxTexture* texture, f32 x, f32 y, f32 width, f32 height, f32 rotation, s8 z)
+	Entity ui_blank_texture(AEGfxTexture* texture, f32 x, f32 y, f32 width, f32 height, f32 rotation, s32 z)
 	{
 		Entity id = ecs.createEntity();
 		//default player values
